Assume the working directory if no arguments passed
Frequently when using prdeps, I want to check the package in the
working directory. However I still have to type it out manually for
prdeps to recognize it.

If a user does not pass any arguments to prdeps, assume they want
to evaluate the current working directory. If you are not in the
GOPATH, we return an error. Add some tests to double check that we are
correctly identifying directories in the GOPATH.